+++ internal/model/interfaces/interfaces.go
