+++ internal/usecase/icon/icon.go
@@ -2,7 +2,7 @@ package icon
 
 import (
 	"github.com/OYE0303/expense-tracker-go/internal/domain"
-	"github.com/OYE0303/expense-tracker-go/internal/usecase/interfaces"
+	"github.com/OYE0303/expense-tracker-go/internal/model/interfaces"
 )
 
 type IconUC struct {
